requirements.txt
flask
llama-index
llama-index-llms-openai
llama-index-llms-groq
tavily-python
python-dotenv 
